Day 23, Part 2
Is finally done! Almost a total re-write to use the A* alogrithm,
which I've now learnt. Thanks Eric!

Code is a bot of a mess towards the end as I was debugging so many
things (see next para) but I cba for now

I misread the conditions and ignored part 3 about
not moving about on the corridor. This made the problem harder
than it needed to be as there was te possibility of blocking
amihpods that needed to move out of the way of each other